Changed tiles from margins (bad for clicking) to borders.
Minor cleanup in the PuzzleCells query.

Thinking about switching to the dice game..., CSV transcripts & trends?
