+++ PuzzleGrid.xaml.cs
@@ -156,9 +156,11 @@ namespace MinesPuzzle
                         Background = PuzzleColors.TileBrush_Unknown,
                         Height = 30,
                         Width = 30,
+                        BorderThickness = new Thickness (1,1,1,1),
+                        //BorderBrush= PuzzleColors.Color_Unknown,
                         /*TODO:  Remove troubleshooting. */
                         //Content = ((int) tag.CellValue).ToString (),
-                        Margin = new Thickness ( 1, 1, 1, 1 ),
+                        //Margin = new Thickness ( 1, 1, 1, 1 ),
                     };
 
                     button.SetValue ( RowProperty, row );
